Bug correction

+++ API/Model/DomainModel/Dish.cs
@@ -1,9 +1,4 @@
 ﻿using MongoDB.Bson.Serialization.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Model.DomainModel
 {
@@ -13,7 +8,7 @@ namespace Model.DomainModel
         public string Id { get; set; }
         public string MenuId { get; set; }
         public string Name { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
     }
 }
